Move StringUtils to ngrinder-core
